приложение работает, методы из репозиториев активны.

+++ src/main/java/com/example/onlinestore/service/AdsService.java
@@ -1,9 +1,10 @@
-package com.example.onlinestore.service.impl;
+package com.example.onlinestore.service;
 
 import com.example.onlinestore.dto.Ads;
 import com.example.onlinestore.dto.CreateAds;
 import com.example.onlinestore.dto.FullAds;
 import com.example.onlinestore.dto.ResponseWrapperAds;
+import com.example.onlinestore.exception.AdsNotFoundException;
 import com.example.onlinestore.mapper.AdsMapper;
 import com.example.onlinestore.model.AdsModel;
 import com.example.onlinestore.model.ImageModel;
@@ -12,31 +13,39 @@ import com.example.onlinestore.repository.AdsRepository;
 import com.example.onlinestore.repository.ImageRepository;
 import com.example.onlinestore.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.awt.*;
 import java.io.IOException;
 import java.util.List;
+import java.util.Optional;
+
 @Slf4j
 @Service
-@Transactional
 public class AdsService {
 
     private final AdsRepository adsRepository;
     private final ImageRepository imageRepository;
+    private final ImageService imageService;
     private final UserRepository userRepository;
     private final UserService userService;
     private final AdsMapper adsMapper;
     private final CommentService commentService;
+    private final PropertyService propertyService;
 
-    public AdsService(AdsRepository adsRepository, ImageRepository imageRepository, UserRepository userRepository, UserService userService, AdsMapper adsMapper, CommentService commentService) {
+    public AdsService(AdsRepository adsRepository, ImageRepository imageRepository, ImageService imageService, UserRepository userRepository, UserService userService, AdsMapper adsMapper, CommentService commentService, PropertyService propertyService) {
         this.adsRepository = adsRepository;
         this.imageRepository = imageRepository;
+        this.imageService = imageService;
         this.userRepository = userRepository;
         this.userService = userService;
         this.adsMapper = adsMapper;
         this.commentService = commentService;
+        this.propertyService = propertyService;
     }
 
     public ResponseWrapperAds getAllAds() {
@@ -53,17 +62,26 @@ public class AdsService {
     }
 
     @Transactional
-    public Ads addAd(CreateAds properties, MultipartFile file) throws IOException {
+    public Ads addAd(CreateAds properties, MultipartFile file,Authentication authentication) throws IOException {
         AdsModel adsModel = adsMapper.toAdsModel(properties);
         ImageModel imageModel = new ImageModel();
 //        imageModel.setFileSize(file.getSize());
         imageModel.setImage(file.getBytes());
         imageRepository.save(imageModel);
         adsModel.setImageModel(imageModel);
-        adsModel.setUserModel(userRepository.findByUsername(userService.getCurrentUsername()).orElseThrow(() -> new UsernameNotFoundException("User not found")));
+        adsModel.setUserModel(userRepository.findByUsername(authentication.getName()).orElseThrow(() -> new UsernameNotFoundException("User not found")));
         adsRepository.save(adsModel);
         return adsMapper.adsModelToAds(adsModel);
     }
+//    public Ads addAd(CreateAds properties, ImageModel image) {
+//        AdsModel adsModel = adsMapper.toAdsModel(properties);
+//        adsModel.setUserModel(userRepository.findByUsername(userService.getCurrentUsername()).orElseThrow(() -> new UsernameNotFoundException("User not found")));
+//        ImageModel imageModel;
+//        imageModel = imageService.save(image);
+//        adsModel.setImageModel(imageModel);
+//        adsRepository.saveAndFlush(adsModel);
+//        return adsMapper.adsModelToAds(adsModel);
+//    }
 
     @Transactional
     public FullAds getAds(Integer id) {
@@ -81,11 +99,26 @@ public class AdsService {
     }
 
     @Transactional
-    public void removeAd(Integer id) {
-        imageRepository.delete(adsRepository.findById(id).map(AdsModel::getImageModel).orElseThrow());
-        commentService.deleteCommentsByAdId(id);
-        adsRepository.deleteById(id);
+    public boolean removeAd(String email, int id) {
+        AdsModel adsModel = adsRepository.findById(id).orElseThrow(AdsNotFoundException::new);
+        UserModel userModel = adsModel.getUserModel();
+//        imageRepository.delete(adsRepository.findById(id).map(AdsModel::getImageModel).orElseThrow());
+//        commentService.deleteCommentsByAdId(id);
+        if (propertyService.isThisUserOrAdmin(email, userModel)) {
+//            try {
+//                Files.deleteIfExists(Path.of(ads.getImage().getFilePath()));
+//            } catch (IOException e) {
+//                throw new RuntimeException(e);
+            adsRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
+//    public void removeAd(Integer id) {
+//        imageRepository.delete(adsRepository.findById(id).map(AdsModel::getImageModel).orElseThrow());
+//        commentService.deleteCommentsByAdId(id);
+//        adsRepository.deleteById(id);
+//    }
 
     @Transactional
     public void updateAdImage(Integer id, MultipartFile file) throws IOException {
@@ -98,14 +131,18 @@ public class AdsService {
     }
 
     @Transactional
-    public Ads updateDto (Integer id, CreateAds properties) {
-        AdsModel adsModel = adsRepository.findById(id).orElseThrow();
-        adsModel.setTitle(properties.getTitle());
-        adsModel.setDescription(properties.getDescription());
-        adsModel.setPrice(properties.getPrice());
-
-        adsRepository.save(adsModel);
-
-        return adsMapper.adsModelToAds(adsModel);
+    public Ads updateDto(Integer id, CreateAds properties, String username) {
+        Optional<AdsModel> optionalAdsModel = adsRepository.findById(id);
+        if (optionalAdsModel.isPresent()) {
+            AdsModel adsModel = optionalAdsModel.get();
+            UserModel userModel = adsModel.getUserModel();
+            if (propertyService.isThisUserOrAdmin(username, userModel)) {
+                adsMapper.toAdsModel(properties);
+                adsRepository.save(adsModel);
+                return adsMapper.adsModelToAds(adsModel);
+            }
+        }
+        return null;
     }
+
 }
